Revised Installation Logs
Changed the update log to a .conf file and added checks to all installation scripts that verify whether that install step has been successfully completed previously.

+++ scripts/install/removes/remove-preinstalled.sh
@@ -18,8 +18,17 @@ source /usr/local/bin/enginescript/scripts/functions/shared/enginescript-common.
 #----------------------------------------------------------------------------------
 # Start Main Script
 
+source /etc/enginescript/install-state.conf
+if [[ "${REMOVES}" = 1 ]]; then
+    echo "REMOVES script has already run"
+    exit 0
+fi
+
 # Remove Previous Installations
 apt-get remove --purge 'apache2.*' mysql-client mysql-common mysql-server 'nginx.*' 'php7\.0.*' 'php7\.1.*' 'php7\.2.*' 'php7\.3.*' 'php7\.4.*' 'php8\.0.*' 'php8\.1.*' -y
 
 # Cleanup
 /usr/local/bin/enginescript/scripts/functions/enginescript-cleanup.sh
+
+# Mark the installation as complete
+echo "REMOVES=1" >> /etc/enginescript/install-state.conf
